flaskr/requirements.txt
arff==0.9
blinker==1.3
cpickle==1.3.3
debugger==1.3
decorator==3.4.0
facebook-sdk==1.0.0a0
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-MongoEngine==0.1.3.dev0
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-Security==1.7.4
Flask-Social==1.6.2
Flask-SocketIO==0.3.7
Flask-Sockets==0.2.0
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
funcsigs==0.4
gevent==1.0.2
gevent-socketio==0.3.6
gevent-websocket==0.9.5
gnureadline==6.3.3
google-api-python-client==1.3.1
greenlet==0.4.9
gunicorn==18.0
httplib2==0.9
ipython==2.1.0
itsdangerous==0.24
Jinja2==2.7.2
MarkupSafe==0.23
matplotlib==1.4.3
mock==1.3.0
mongoengine==0.9.0
networkx==1.9.1
nltk==3.1
nose==1.3.7
numpy==1.10.0
oauth2==1.5.211
oauth2client==1.4.5
passlib==1.6.2
pbr==1.8.0
pyasn1==0.1.7
pyasn1-modules==0.0.5
PyAudio==0.2.9
pydot==1.0.2
pyechonest==9.0.0
pymongo==3.0
pyo==0.7.5
pyparsing==1.5.7
pyreadline==2.0
python-dateutil==2.4.0
python-engineio==0.7.2
python-gflags==2.0
python-graph-core==1.8.2
python-graph-dot==1.8.2
python-socketio==0.6.1
pytz==2015.7
redis==2.10.5
requests==2.6.0
rsa==3.1.4
simplejson==3.6.5
six==1.9.0
SpeechRecognition==3.2.1
spotipy==2.3.3
SQLAlchemy==0.9.9
uritemplate==0.6
vboxapi==1.0
virtualenv==12.1.1
Werkzeug==0.9.4
WTForms==2.0.2
wxPython==3.0.2.0
wxPython-common==3.0.2.0
